Run contributed tasks automatically if configured in the workspace.
